requirements.txt
pandas==2.1.1
numpy==1.26.0
dash==2.13.0
dash-bootstrap-components==1.5.0
dash-iconify==0.1.2
dash-mantine-components==0.12.1
Pillow==10.0.1
tensorflow==2.14.0
nbformat==5.9.2